renamed user card to search card
