+++ src/components/ApplicationViews.js
@@ -8,7 +8,6 @@ import useSimpleAuth from '../hooks/ui/useSimpleAuth'
 import Home from './home/Home'
 import DesignForm from './design/DesignForm'
 
-import UserSearch from './user/UserSearch'
 import UserHome from './user/UserHome'
 
 
@@ -47,12 +46,6 @@ const ApplicationViews = props => {
                     return <DesignForm {...props}/>
                 }
             }} />
-
-            <Route path="/user/search" render={props => {
-                if(isAuthenticated()) {
-                    return <UserSearch />
-                }
-            }}/>
 
             <Route path="/following/" render={props => {
                 if(isAuthenticated()) {
